requirements.txt
# PyTorch with CUDA support (CUDA 11.8 or 12.1)
# For CUDA 11.8: --index-url https://download.pytorch.org/whl/cu118
# For CUDA 12.1: --index-url https://download.pytorch.org/whl/cu121
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
timm>=0.9.0
scipy>=1.11.0
networkx>=3.1
seaborn>=0.12.0
tqdm>=4.66.0
pillow>=10.0.0
scikit-image>=0.21.0
matplotlib>=3.7.0
scanpy>=1.9.0
anndata>=0.9.0
h5py>=3.9.0
numba>=0.57.0
limix>=3.0.0
statsmodels>=0.14.0
lifelines>=0.27.0
scikit-survival>=0.21.0